requirements.txt
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1.1
django-mptt==0.8.7
psycopg2==2.6.1
olefile==0.44
Pillow==4.0.0
rcssmin==1.0.6
rjsmin==1.0.12
sorl-thumbnail==12.3
django-extensions==1.7.6
django-ckeditor==5.2.1
djangorestframework==3.5.3



